Stop using `echo -n` as it may cause Windows issues
It's possible that Windows is doing something strange here that's
causing the `echo` tobe handled by the Windows shell instead of as part
of the kubectl input passed to `sh`. Switching to plain echo and then
stripping newlines from the result should produce the same result
without any cross-platform hiccups.